+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add user id and IP address to rollbar server side error logging [#841](https://github.com/PublicMapping/districtbuilder/pull/841)
 - Add majority race metric to project sidebar [#853](https://github.com/PublicMapping/districtbuilder/pull/853)
 - Add additional keyboard shortcuts [#854](https://github.com/PublicMapping/districtbuilder/pull/854)
+- Add configurable slider to increase size of paintbrush selection tool [#835](https://github.com/PublicMapping/districtbuilder/pull/835)
+
 
 ### Changed